# server/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Phantoma Server API
  description: Backend API for Phantoma - Server Reversing & Hacking tool
  version: "1.0.0"

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      tags: [system]
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/nmap/scan:
    post:
      summary: Run an nmap network scan
      tags: [nmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
            example:
              target: "scanme.nmap.org"
              flags: ["-sV", "-p", "80,443"]
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /api/v1/nikto/scan:
    post:
      summary: Run a Nikto web vulnerability scan
      tags: [nikto]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
            example:
              target: "http://example.com"
              flags: ["-Tuning", "1"]
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

components:
  schemas:
    ScanRequest:
      type: object
      required: [target]
      properties:
        target:
          type: string
          description: IP address, hostname, or URL to scan
        flags:
          type: array
          items:
            type: string
          description: Optional tool-specific flags

    ScanResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ScanResult"

    ScanResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        error:
          type: string

    HealthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            status:
              type: string
            service:
              type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string